requirements.txt
Flask==2.3.3
gspread==5.12.0
google-auth==2.29.0
